style: overhauled CSS parser tests to be more readable
and consolidated the XPathVisitor tests from css/test_parser.rb into
css/test_xpath_visitor.rb so we're testing xpath visitor behavior in
one place.
